# openapi.yaml
openapi: 3.1.0
info:
  title: AetherPro Omni Voice Gateway
  version: 1.0.0
  description: >
    Multimodal voice gateway providing real-time audio-to-audio interaction,
    read-aloud TTS, transcription, and model introspection.
servers:
  - url: http://localhost:8004
    description: External host port
  - url: http://voice-gateway:8000
    description: Internal Docker network
paths:
...
  /voice/batch:
    post:
      summary: Batch text-to-speech
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
      responses:
        "200":
          description: Audio output
          content:
            audio/wav:
              schema:
                type: string
                format: binary
...
  /voice/transcribe:
    post:
      summary: Audio transcription
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string }
                  confidence: { type: number }
...